+++ modules/core/karrio/server/core/utils.py
@@ -159,6 +159,7 @@ def error_wrapper(func):
     return wrapper
 
 
+
 def async_wrapper(func):
     @functools.wraps(func)
     def wrapper(*args, run_synchronous: bool = False, **kwargs):
@@ -359,6 +360,32 @@ def post_processing(methods: List[str] = None):
                     processes,
                     result,
                 )
+
+            setattr(klass, name, wrapper)
+
+        return klass
+
+    return class_wrapper
+
+
+def pre_processing(methods: List[str] = None):
+    def class_wrapper(klass):
+        setattr(
+            klass,
+            "pre_process_functions",
+            getattr(klass, "pre_process_functions", None) or [],
+        )
+
+        for name in methods:
+            method = getattr(klass, name)
+
+            def wrapper(*args, _method=method, **kwargs):
+                context = kwargs.get("context")
+
+                for process in klass.pre_process_functions:
+                    process(context, **kwargs)
+
+                return _method(*args, **kwargs)
 
             setattr(klass, name, wrapper)